+++ pdata/internal/common.go
@@ -262,7 +262,16 @@ func (v Value) SliceVal() Slice {
 // If the Type() is not ValueTypeBytes then returns false.
 // Calling this function on zero-initialized Value will cause a panic.
 // Modifying the returned []byte in-place is forbidden.
+// Deprecated: [0.51.0] Use MBytesVal instead.
 func (v Value) BytesVal() []byte {
+	return v.orig.GetBytesValue()
+}
+
+// MBytesVal returns the []byte value associated with this Value.
+// If the Type() is not ValueTypeBytes then returns false.
+// Calling this function on zero-initialized Value will cause a panic.
+// Modifying the returned []byte in-place is forbidden.
+func (v Value) MBytesVal() []byte {
 	return v.orig.GetBytesValue()
 }
 
@@ -299,7 +308,17 @@ func (v Value) SetBoolVal(bv bool) {
 // Calling this function on zero-initialized Value will cause a panic.
 // The caller must ensure the []byte passed in is not modified after the call is made, sharing the data
 // across multiple attributes is forbidden.
+// Deprecated: [0.51.0] Use SetMBytesVal instead.
 func (v Value) SetBytesVal(bv []byte) {
+	v.orig.Value = &otlpcommon.AnyValue_BytesValue{BytesValue: bv}
+}
+
+// SetMBytesVal replaces the []byte value associated with this Value,
+// it also changes the type to be ValueTypeBytes.
+// Calling this function on zero-initialized Value will cause a panic.
+// The caller must ensure the []byte passed in is not modified after the call is made, sharing the data
+// across multiple attributes is forbidden.
+func (v Value) SetMBytesVal(bv []byte) {
 	v.orig.Value = &otlpcommon.AnyValue_BytesValue{BytesValue: bv}
 }
 
@@ -709,7 +728,18 @@ func (m Map) InsertBool(k string, v bool) {
 // No action is applied to the map where the key already exists.
 // The caller must ensure the []byte passed in is not modified after the call is made, sharing the data
 // across multiple attributes is forbidden.
+// Deprecated: [0.51.0] Use InsertMBytes instead.
 func (m Map) InsertBytes(k string, v []byte) {
+	if _, existing := m.Get(k); !existing {
+		*m.orig = append(*m.orig, newAttributeKeyValueBytes(k, v))
+	}
+}
+
+// InsertMBytes adds the []byte Value to the map when the key does not exist.
+// No action is applied to the map where the key already exists.
+// The caller must ensure the []byte passed in is not modified after the call is made, sharing the data
+// across multiple attributes is forbidden.
+func (m Map) InsertMBytes(k string, v []byte) {
 	if _, existing := m.Get(k); !existing {
 		*m.orig = append(*m.orig, newAttributeKeyValueBytes(k, v))
 	}
@@ -764,9 +794,20 @@ func (m Map) UpdateBool(k string, v bool) {
 // No action is applied to the map where the key does not exist.
 // The caller must ensure the []byte passed in is not modified after the call is made, sharing the data
 // across multiple attributes is forbidden.
+// Deprecated: [0.51.0] Use UpdateMBytes instead.
 func (m Map) UpdateBytes(k string, v []byte) {
 	if av, existing := m.Get(k); existing {
-		av.SetBytesVal(v)
+		av.SetMBytesVal(v)
+	}
+}
+
+// UpdateMBytes updates an existing []byte Value with a value.
+// No action is applied to the map where the key does not exist.
+// The caller must ensure the []byte passed in is not modified after the call is made, sharing the data
+// across multiple attributes is forbidden.
+func (m Map) UpdateMBytes(k string, v []byte) {
+	if av, existing := m.Get(k); existing {
+		av.SetMBytesVal(v)
 	}
 }
 
@@ -835,9 +876,23 @@ func (m Map) UpsertBool(k string, v bool) {
 // updated to the map where the key already existed.
 // The caller must ensure the []byte passed in is not modified after the call is made, sharing the data
 // across multiple attributes is forbidden.
+// Deprecated: [0.51.0] Use UpsertMBytes instead.
 func (m Map) UpsertBytes(k string, v []byte) {
 	if av, existing := m.Get(k); existing {
-		av.SetBytesVal(v)
+		av.SetMBytesVal(v)
+	} else {
+		*m.orig = append(*m.orig, newAttributeKeyValueBytes(k, v))
+	}
+}
+
+// UpsertMBytes performs the Insert or Update action. The []byte Value is
+// inserted to the map that did not originally have the key. The key/value is
+// updated to the map where the key already existed.
+// The caller must ensure the []byte passed in is not modified after the call is made, sharing the data
+// across multiple attributes is forbidden.
+func (m Map) UpsertMBytes(k string, v []byte) {
+	if av, existing := m.Get(k); existing {
+		av.SetMBytesVal(v)
 	} else {
 		*m.orig = append(*m.orig, newAttributeKeyValueBytes(k, v))
 	}